refactor(fipe): refactore automakers and reference tables
