+++ pages/api/fipe/marcas/v1/carros.js
@@ -1,10 +1,8 @@
 import app from '@/app';
 import BadRequestError from '@/errors/bad-request';
 
-import {
-  listCarAutomakers,
-  listReferenceTables,
-} from '@/services/fipe/automakers';
+import { listCarAutomakers } from '@/services/fipe/automakers';
+import { listReferenceTables } from '@/services/fipe/referenceTable';
 
 async function FipeCarsAutomakers(request, response) {
   const referenceTableCode = request.query.tabela_referencia;
