+++ services/fipe/constants.js
@@ -1,0 +1,7 @@
+export const VEHICLE_TYPE = {
+  CAR: 1,
+  MOTORCYCLE: 2,
+  TRUCK: 3,
+};
+
+export const FIPE_URL = 'https://veiculos.fipe.org.br/api';
